+++ src/client/games/slashchain/client.tsx
